[#855] Add check if the Alias is already used
Closes #855

- Added `UsedAliasesChecker` utility which is used to register aliases
  that are in use.
- `prepare()` and `wipe()` methods now check if the current alias can't
  be used and if not then `InitializerError.aliasAlreadyInUse` is
  thrown/emitted.
- Some public methods that could cause harm if used with the Alias that
  is already in use now throw `SynchronizerError.notPrepared`. Thanks to
  this the client app is forced to call `prepare()` first. And
  `prepare()` does check for the Alias.
- Added tests for new conditions.

+++ Sources/ZcashLightClientKit/Block/FilesystemStorage/FSCompactBlockRepository.swift
@@ -126,7 +126,9 @@ extension FSCompactBlockRepository: CompactBlockRepository {
 
     func clear() async throws {
         try await Task {
-            try self.fileManager.removeItem(at: self.fsBlockDbRoot)
+            if self.fileManager.fileExists(atPath: self.fsBlockDbRoot.path) {
+                try self.fileManager.removeItem(at: self.fsBlockDbRoot)
+            }
             try create()
         }.value
     }
